BE: "chat_ws_test_1" has been renamed to "chat_ws_test_base".
